Update constructor for uv_buf_t

+++ cmod/uvwrap/src/uv-common.c
@@ -139,7 +139,7 @@ size_t uv_loop_size(void) {
   return sizeof(uv_loop_t);
 }
 
-uv_buf_t uv_buf_init(char* base, unsigned int len) {
+uv_buf_t uv_buf_init(char* base, size_t len) {
   uv_buf_t buf;
   buf.base = base;
   buf.len = len;
